refactor: 移动到 pipeline.py 中
