+++ maa_mcp/core.py
@@ -162,6 +162,8 @@ mcp = FastMCP(
        - 反复执行"运行→分析→修复→运行"循环
        - 直到 Pipeline 稳定运行成功
        - 建议多次验证以确保鲁棒性
+    
+    5. 通过浏览器打开最终版本的 Pipeline 可视化界面供用户查看和保存
     """,
 )
