+++ maa_mcp/pipeline.py
@@ -8,9 +8,11 @@ Pipeline 生成支持模块
 """
 
 import json
+import webbrowser
 from datetime import datetime
 from pathlib import Path
 from typing import Optional
+from lzstring import LZString
 
 from maa.tasker import TaskDetail
 
@@ -236,6 +238,20 @@ Pipeline 是 MaaFramework 的任务流水线，采用 JSON 格式描述，由若
 '''
 
 
+"""
+MPE 相关配置
+"""
+
+# MPE 分享协议版本
+MPE_SHARE_VERSION = 1
+# URL 参数名
+MPE_SHARE_PARAM = "shared"
+# 默认 MPE 基准地址
+MPE_BASE_URL = "https://mpe.codax.site/stable"
+# URL 最大大小限制
+MPE_MAX_URL_SIZE = 60 * 1024  # 60KB
+
+
 @mcp.tool(
     name="get_pipeline_protocol",
     description="""
@@ -445,3 +461,67 @@ def run_pipeline(
         return "任务执行失败，无法获取执行详情"
 
     return task_detail
+
+
+
+
+def generate_share_link(
+    pipeline_obj: dict
+) -> str:
+    # 生成分享链接
+    payload = {
+        "v": MPE_SHARE_VERSION,
+        "d": pipeline_obj,
+    }
+    json_string = json.dumps(payload, ensure_ascii=False, separators=(",", ":"))
+    lz = LZString()
+    compressed = lz.compressToEncodedURIComponent(json_string)
+    share_url = f"{MPE_BASE_URL}?{MPE_SHARE_PARAM}={compressed}"
+    return share_url
+
+
+@mcp.tool(
+    name="open_pipeline_in_browser",
+    description="""
+    通过浏览器打开 Pipeline JSON 可视化界面。
+
+    参数：
+    - pipeline_file_path: Pipeline JSON 文件的本地路径（字符串）
+
+    功能说明：
+    该工具会读取指定路径的 Pipeline JSON 文件，将数据压缩编码后生成一个分享链接，
+    并自动在系统默认浏览器中打开，方便用户可视化查看工作流结构。
+
+    注意：
+    - 此工具无返回值，仅执行打开浏览器的操作
+    - 仅在用户要求查看 Pipeline 可视化流程图时使用
+    - 传入的文件路径必须指向一个有效的本地 JSON 文件
+    - 如果生成的 URL 超过 60KB，将返回错误提示而不打开浏览器
+    """,
+)
+def open_pipeline_in_browser(
+    pipeline_file_path: str
+) -> None:
+    # 读取文件内容
+    file_path = Path(pipeline_file_path)
+
+    if not file_path.exists():
+        raise FileNotFoundError(f"Pipeline 文件不存在: {pipeline_file_path}")
+    if not file_path.is_file():
+        raise ValueError(f"路径不是文件: {pipeline_file_path}")
+
+    with open(file_path, "r", encoding="utf-8") as f:
+        pipeline_obj = json.load(f)
+
+    # 生成分享链接
+    share_url = generate_share_link(pipeline_obj, base_url)
+
+    # 检查 URL 大小
+    url_size = len(share_url.encode("utf-8"))
+    if url_size > MPE_MAX_URL_SIZE:
+        size_kb = url_size / 1024
+        raise ValueError(
+            f"生成的分享链接过大（{size_kb:.2f} KB），请自行通过复制或文件的方式导入 Pipeline 至 MPE。"
+        )
+
+    webbrowser.open(share_url)
